Expand environment out beyond the size of the track
because I wasn't having any problems with DynamicTerrain or rendering quads, but that the box I had set the sky onto was too small to cover the whole track. So I switched the sky onto a much larger sphere to give us a better viewing experience, all the way out to our horizon.

Plus this cool camera animation flying in a big circle around the track is nice